django_taxi_analysis/requirements.txt
Django==4.2.7
mysql-connector-python==8.2.0
pandas==1.5.3
numpy==1.24.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8 
scikit-learn==1.7.0 
requests 